Move manual hook handler to a separate header
